textrender: Rewire visual bell
We want to split the terminal into a reusable component, longer term. To
do this, we need to - longer term - move util-type code to be used from
the application, not the terminal widget.

So: here, we emit visualBell unconditionally from TextRender, and decide
how to handle that on the application side.