Improve the calculation/balance of split sizes
The previous code had a 'widthRatio' for each child in the split, and
they were expected to sum to 1 for all children. Adding a new child
would halve the ratio on the active child, which was awkward.

Instead, calculate sizes based on the sum of sizeRatio for all children,
so that each child receives an amount proportional to its sizeRatio.
Also, take the size of splitters into account for this calculation.

The widthRatio and heightRatio properties are combined into a single
sizeRatio, because an item can only have one value for the direction of
it's parent splitter.