Correct the 256 color table
The table for 256 colors was slightly off from every other reference I
can find. To better work with color schemes, this uses exactly the
normal xterm colors.

Also, don't increment the color by 8 for boldness unless it is one of
the first 16; incrementing by 8 in the 256 color space is not at all
what you'd want.

https://jonasjacek.github.io/colors/ is a useful resource for the xterm
256 color table; other sources agree on the same set of colors.