literm: instantly update when data arrives
otherwise, in a worst case, we might miss the frame deadline when
updating in terminal, and miss it again due to the polish delay, which
can lead to quite a bit of latency.

the timers are now purely for debouncing rather than delaying: we
immediately process, but delay future processing for a while.