terminal: Improve SGR handling correctness and understandability
I have no idea what the params.count - pidx >= 5 check was attempting
to do, but I think it's wrong:

When we enter this code, pidx is pointing at p, which is 38 or 48.
Following that, we have a color type (5, in the case of 16 bit),
plus 3 colors. That should mean that we want 4 parameters
(the 5, and 3 colors), not 5 -- but rather than perpetrate the mess,
I think it's better to just improve the quality of the parsing.

Also added a test script for 16bit color testing, hauled off of the internets.