Clean up conditional handling a little bit
The only two targets that are (really) used at all are desktop and nemo (/Jolla)
So let's have a single way to pick which of them we are building for: LITERM_TARGET.
This file is included, and is expected to set up everything target-specific.

It's done as an include so that someone can still theoretically write
their own .pri to turn off e.g. feedback if they want.